extract useSectionY into shared hook
